extract remote api typings
