+++ figma-remote-types/README.md
@@ -1,0 +1,22 @@
+# `Figma-Remote-Types` _(`@design-sdk/figma-remote-types`)_
+
+> Type definitions of figma remote api (figma rest api)
+
+## Installation
+
+```sh
+yarn add @design-sdk/figma-remote-types
+
+# or with npm
+npm i @design-sdk/figma-remote-types
+```
+
+## Usage: Importing types
+
+```ts
+import { Paint } from "@design-sdk/figma-remote-types";
+```
+
+## Disclamer
+
+The figma remote api responce type definitions from "[figma-js](https://github.com/jongold/figma-js/blob/master/src/figmaTypes.ts)" (MIT)
