--- figma-remote/lib/api/README.md
@@ -1,1 +1,0 @@
-# The figma remote api responce type definitions from "[figma-js](https://github.com/jongold/figma-js/blob/master/src/figmaTypes.ts)" (MIT)
