+++ figma-remote/lib/converters/constraints-mixin.convert.ts
@@ -1,5 +1,5 @@
 import { ConstraintMixin } from "@design-sdk/figma-types";
-import { LayoutConstraint } from "../types";
+import { LayoutConstraint } from "@design-sdk/figma-remote-types";
 import { convertFigmaRemoteLayoutConstraintsToFigmaConstraints } from "./layout-constraints.convert";
 export function convertForFigmaConstraintsMixin(
   _do: LayoutConstraint,
