+++ figma-remote/lib/index.ts
@@ -1,4 +1,4 @@
 export * as api from "./api";
-export * as types from "./types";
+export * as types from "@design-sdk/figma-remote-types";
 export * as mapper from "./mapper";
 export * as fetch from "./fetch";
