+++ figma-remote/lib/mapper/group.mapper.ts
@@ -1,10 +1,6 @@
 import { Figma } from "@design-sdk/figma-types";
-import {
-  blendBaseNode,
-  blendCornerNode,
-  blendRectangleCornerNode,
-} from "../blenders";
-import { Group } from "../types";
+import { blendBaseNode } from "../blenders";
+import { Group } from "@design-sdk/figma-remote-types";
 import { MappingGroupNode } from "./mapping-instance";
 
 export function mapFigmaRemoteGroupToFigma(remGroup: Group): Figma.GroupNode {
