+++ figma-remote/package.json
@@ -23,7 +23,9 @@
     "publishConfig": {
         "access": "public"
     },
+    "dependencies": {},
     "devDependencies": {
-        "typescript": "^4.3.5"
+        "typescript": "^4.3.5",
+        "@design-sdk/figma-remote-types": "0.0.0"
     }
 }
